+++ src/pages/post/PostDetailPage.jsx
@@ -1,15 +1,15 @@
 import { useParams, useRouteLoaderData } from "react-router-dom";
 import { Helmet } from "react-helmet-async";
-import useQuestions from "./components/useQuestions";
-import useSubject from "./components/useSubject";
+import useQuestions from "./hooks/useQuestions";
+import useSubject from "./hooks/useSubject";
 import {
   QuestionForm,
   FeedList,
   FeedListHeader,
   FeedCardList,
   FeedCard,
-  Message,
 } from "@components/FeedCard";
+import { Message } from "./components/Message";
 
 export default function PostDetailPage() {
   const { id } = useParams();
